+++ packages/cfd/src/Containers/cfd-password-modal/handle-password-input-change.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { FormikErrors } from 'formik';
+
+import { TCFDPasswordFormValues } from './cfd-password-modal.types';
+
+const handlePasswordInputChange = (
+    e: React.ChangeEvent<HTMLInputElement>,
+    handleChange: (el: React.ChangeEvent<HTMLInputElement>) => void,
+    validateForm: (values?: TCFDPasswordFormValues) => Promise<FormikErrors<TCFDPasswordFormValues>>,
+    setFieldTouched: (field: string, isTouched?: boolean, shouldValidate?: boolean) => void
+) => {
+    handleChange(e);
+    validateForm().then(() => {
+        setFieldTouched('password', true);
+    });
+};
+
+export default handlePasswordInputChange;
